Commentation on scripts
Added comments to Triggers and Verbs. Changed the named of certain Triggers that were toSomething to whenSomething.
