+++ Verbs/toEnable.cs
@@ -4,16 +4,24 @@ using UnityEngine;
 
 public class toEnable : Verb
 {
-    public Verb[] triggeredVerbs;
+    /*
+     * This verb will make the target object active.
+     */
+
+    //______Variable Declarations_____________________
     public GameObject target;
+    //________________________________________________
+    public Verb[] triggeredVerbs;
 
     private void Awake()
     {
         SetAudio();
     }
 
+    // Update is called once per frame
     private void Update()
     {
+        //Checks the active state of the object. If active will set the target object to active.
         if (isActive)
         {
             EndVerb();
