+++ Verbs/toLoad.cs
@@ -5,8 +5,15 @@ using UnityEngine.SceneManagement;
 
 public class toLoad : Verb {
 
-    public string nextScene;
+    /*
+     * This verb will load a new scene. Scene is provided by user.
+     */
 
+    //______Variable Declarations_____________________
+    public string nextScene;
+    //________________________________________________
+
+    // Update is called once per frame
     void Update()
     {
         if (isActive)
