+++ Verbs/toSequence.cs
@@ -4,9 +4,17 @@ using UnityEngine;
 
 public class toSequence : Verb
 {
-    public bool randomize;
+    /* 
+     * 
+     */
 
-	[SerializeField]
+    //______Variable Declarations_____________________
+    //a variable that works like a true/false statement to determine if sequence shoudl be random.
+    public bool randomize;
+    //_________________________________________________
+
+    //SerializeField is just making this private variable, below, visible in the Unity Editor.
+    [SerializeField]
 	private int currentNumber = 0;
 
 	[SerializeField]
@@ -18,8 +26,11 @@ public class toSequence : Verb
 	void Awake () 
 	{
 		SetAudio();
+        //checks the verb toSequence() to check how many trigger verbs are currently in the list.
 		numberOfSteps = triggeredVerbs.Length;
 
+        //if boolean is true, the starting point for the sequence will be set to random between 0
+        //and the total number of triggeredVerbs
 		if(randomize)
 		{
 			currentNumber = Random.Range(0, numberOfSteps);
@@ -32,6 +43,7 @@ public class toSequence : Verb
             PlayAudio();
     }
 
+    // FixedUpdate is called once per frame
     void FixedUpdate () {
 		if(isActive && numberOfSteps != 0)
 		{
@@ -40,7 +52,9 @@ public class toSequence : Verb
 
             if(randomize){
             	int next = Random.Range(0, numberOfSteps);
-				currentNumber = (next == currentNumber)? ((next + 1) % numberOfSteps) : next;
+                // the ? and : in this statement are part of a ternary operator
+                // Condition? If true : If false
+                currentNumber = (next == currentNumber)? ((next + 1) % numberOfSteps) : next;
 			}
 			else{
 				currentNumber = (currentNumber + 1) % numberOfSteps;
