+++ Verbs/toSpawn.cs
@@ -4,10 +4,18 @@ using UnityEngine;
 
 public class toSpawn : Verb
 {
+    /*
+     * Object will spawn, or be created, at a target location(spawnPoint)
+     * User provides the type of object that spawns(theCreated) as well as the target location.
+     * User can also decide if the Object the verb is on will destroy at the end of this verb.
+     */
 
+    //______Variable Declarations_____________________
     public GameObject theCreated;
     public Vector3 spawnPoint;
+    //a variable that works like a true/false statement to determine destroying Object. 
     public bool destroyOnBirth;
+    //________________________________________________
     public Verb[] triggeredVerbs;
 
     void Awake()
@@ -23,10 +31,13 @@ public class toSpawn : Verb
             PlayAudio();
     }
 
+    // Update is called once per frame
     private void Update()
     {
         if (isActive)
         {
+            //Creates a clone of the object provided at the location provided.
+            //Quaternion is just setting the rotation of the object
             Instantiate(theCreated, spawnPoint, Quaternion.identity);
 
             if (destroyOnBirth)
