+++ Verbs/toStop.cs
@@ -3,6 +3,10 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class toStop : Verb {
+    /*
+     * This verb will stop all actions on the object. It will stop
+     * all active triggerverbs
+     */
 
     public Verb[] triggeredVerbs;
