+++ Verbs/toTeleport.cs
@@ -4,8 +4,17 @@ using UnityEngine;
 
 public class toTeleport : Verb {
 
-	public Vector3 targetPos;
+    /* 
+     * Object will instantly move its position to a target position. The user can decided
+     * if this movement will be two-way and teleport back to a second position.
+     * User provides target position(targetPos) and a second position(pastPos)
+     */
+
+    //______Variable Declarations_____________________
+    public Vector3 targetPos;
 	public Vector3 pastPos;
+    //a variable that works like a true/false statement to determine two-way teleporting. 
+    //________________________________________________
     public bool twoWayTeleporting;
 
     public Verb[] triggeredVerbs;
@@ -21,13 +30,18 @@ public class toTeleport : Verb {
             PlayAudio();
     }
 
+    // Update is called once per frame
     void Update () {
 		if(isActive)
 		{
+            //Two conditions that check if the boolean above way set to true/false
             if (twoWayTeleporting)
             {
+                //sets pastPost to the current position of the Object
                 pastPos = this.transform.position;
+                //sets the current position of Object to target position, ie. Teleports
                 this.transform.position = targetPos;
+                //switches location again before it will repeat at the start of this condition.
                 targetPos = pastPos;
             }
             else
