+++ Verbs/toWait.cs
@@ -4,7 +4,14 @@ using UnityEngine;
 
 public class toWait : Verb {
 
-	public float duration = 3.0f;
+    /*
+     * Object will stay and wait for a duration of time
+     * The user can provide the duration or it will be set to 3 as default.
+     */
+
+    //______Variable Declarations_____________________
+    public float duration = 3.0f;
+    //________________________________________________
 	public Verb[] triggeredVerbs;
 
 	private float timePassed;
@@ -24,6 +31,7 @@ public class toWait : Verb {
     void Update () {
 		if(isActive)
 		{
+            //Time.deltaTime will increase the timePassed variable by time units instead of frames.
 			timePassed += Time.deltaTime;
 
 			if(timePassed >= duration)
@@ -35,6 +43,7 @@ public class toWait : Verb {
 		}
 	}
 
+    //overrides the Conjugate() function in the Verb() class.
 	override public void Conjugate () {
 		timePassed = 0.0f;
         base.Conjugate();
